refactor(Section): 구간 관련 Section -> Line으로 이동
구간 관련 Section -> Line으로 이동
- Section 관련 테스트 -> LineAcceptanceTest 에 Nested로 추가
- 나머지는 모두 이동만 함
