+++ src/main/java/nextstep/subway/line/domain/Distance.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section.domain;
+package nextstep.subway.line.domain;
 
 import nextstep.subway.common.exception.NegativeNumberDistanceException;
