+++ src/main/java/nextstep/subway/line/domain/Section.java
@@ -1,7 +1,6 @@
-package nextstep.subway.section.domain;
+package nextstep.subway.line.domain;
 
 import nextstep.subway.common.exception.SameStationsInSectionException;
-import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.Embedded;
