+++ src/test/java/nextstep/subway/line/LineAcceptanceFixture.java
@@ -24,6 +24,14 @@ public class LineAcceptanceFixture {
         return params;
     }
 
+    public static Map<String, String> createParams(StationResponse upStation, StationResponse downStation, int distance) {
+        Map<String, String> params = new HashMap<>();
+        params.put("upStationId", upStation.getId().toString());
+        params.put("downStationId", downStation.getId().toString());
+        params.put("distance", String.valueOf(distance));
+        return params;
+    }
+
     public static Map<String, String> createParams(String name, String color, Long upStationId, Long downStationId, int distance) {
         Map<String, String> params = new HashMap<>();
         params.put("name", name);
@@ -77,6 +85,23 @@ public class LineAcceptanceFixture {
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().delete("/lines/" + id)
+                .then().log().all().extract();
+    }
+
+    public static ExtractableResponse<Response> 구간_추가를_요청한다(Long lineId, StationResponse upStation, StationResponse downStation, int distance) {
+        Map<String, String> params = createParams(upStation, downStation, distance);
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/lines/" + lineId.toString() + "/sections")
+                .then().log().all().extract();
+    }
+
+    public static ExtractableResponse<Response> 구간_삭제를_요청한다(Long lineId, Long stationId) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .param("stationId", stationId)
+                .when().delete("/lines/" + lineId.toString() + "/sections")
                 .then().log().all().extract();
     }
