+++ src/test/java/nextstep/subway/line/domain/SectionTest.java
@@ -1,8 +1,10 @@
-package nextstep.subway.section.domain;
+package nextstep.subway.line.domain;
 
 import nextstep.subway.common.exception.NegativeNumberDistanceException;
 import nextstep.subway.common.exception.SameStationsInSectionException;
+import nextstep.subway.line.domain.Distance;
 import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
